Add TimeManager class to manage timers, elapsed time, frame count, etc

+++ GameObject.cpp
@@ -1,8 +1,8 @@
 #include "GameObject.h"
 
 // TODO: lots of parameters and long initializer list, maybe create Time object?
-GameObject::GameObject(MeshBase& mesh, ShaderProgram& shaderProgram, glm::mat4 transform, float& elapsedTime,
-                       float& deltaTime, TextureManager& texman)
+GameObject::GameObject(MeshBase& mesh, ShaderProgram& shaderProgram, glm::mat4 transform, double& elapsedTime,
+                       double& deltaTime, TextureManager& texman)
     : mesh(mesh), transform(transform), elapsedTime(elapsedTime), deltaTime(deltaTime), texman(texman),
       shaderProgram(shaderProgram) {
     glGenVertexArrays(1, &vao);
