+++ include/CubeObject.h
@@ -4,7 +4,8 @@
 class CubeObject : public GameObject {
   public:
     float RotSpeed = 0;
-    CubeObject(MeshBase& mesh, ShaderProgram& shaderProgram, glm::mat4 transform, float& elapsedTime, float& deltaTime,
+    CubeObject(MeshBase& mesh, ShaderProgram& shaderProgram, glm::mat4 transform, double& elapsedTime,
+               double& deltaTime,
                TextureManager& texman);
     void Tick() override;
     void Draw(const Camera& camera) const override;
